requirements.txt
# ──────────────────────────────────────────────
#  API framework
# ──────────────────────────────────────────────
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
python-multipart==0.0.22      # ⚠️  REQUIRED — FastAPI will crash at startup without this
starlette==0.52.1

# ──────────────────────────────────────────────
#  Deep learning
#  Pinned to versions that work with facenet-pytorch 2.6.0
# ──────────────────────────────────────────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
torchvision==0.17.2+cpu

# ──────────────────────────────────────────────
#  Computer vision & face models
# ──────────────────────────────────────────────
opencv-python-headless==4.10.0.84   # last version compatible with numpy 1.x
Pillow==10.2.0
facenet-pytorch==2.6.0
ultralytics==8.4.21

# ──────────────────────────────────────────────
#  Numerical
# ──────────────────────────────────────────────
numpy==1.26.4

# ──────────────────────────────────────────────
#  Monitoring
# ──────────────────────────────────────────────
prometheus-fastapi-instrumentator==6.1.0